Reorganize PubSub modules into domain-specific namespaces
